UI update. Needs cleanu

+++ Dylan/Models/Performances/sPerformance.swift
@@ -12,7 +12,7 @@ struct sPerformance: Codable {
     let venue: String
     let songs: [sSong]
     let date: Double?
-    var lbNumbers: [Double]?
+    var lbNumbers: [Int]?
     
     var id: String {
         venue + String(date ?? 0)
